move all files into root directory
