+++ create_basic_characteristics_table.py
