+++ slick.sh
